% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addPlateLayout}
\alias{addPlateLayout}
\title{Adding Plate Layout Information to reader data frame}
\usage{
addPlateLayout(reader_df, barcode = "0000", setValueNA = FALSE,
  setSlopeNA = FALSE)
}
\arguments{
\item{reader_df,}{barcode="0000", setValueNA=TRUE, setSlopeNA=FALSE}
}
\description{
addPlateLayout
}
\note{
todo : merging bug with multiple measurements per data frame (e.g. groth data)
}
\examples{
addPlateLayout(reader_df, barcode="0000", setValueNA=TRUE, setSlopeNA=FALSE)
}
\keyword{plate}
\keyword{readers}

