% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genWellNumbers}
\alias{genWellNumbers}
\title{Well Number Generator}
\usage{
genWellNumbers(padding = 2)
}
\arguments{
\item{padding}{}
}
\value{
factors of well numbers
}
\description{
Generator for microtiter plate well coordinates
}
\details{
genWellNumbers
}
\examples{
genWellNumbers(padding=1)
   ->  A1, A2, ...., H12
  genWellNumbers(padding=3)
   -> A001, A002, ..., H012
}
\keyword{plate}
\keyword{readers}

