% Generated by roxygen2 (4.0.2): do not edit by hand
\name{importReaderData}
\alias{importReaderData}
\title{Importing Data from Microtiter Plate Readers}
\usage{
importReaderData(filename = "", path = ".", device = "omega",
  method = "SPabs", wavelength = 0, barcode = "0000", layout = TRUE,
  PLC = FALSE)
}
\arguments{
\item{path}{(string)}

\item{device}{(string) - one of the above mentioned devices}

\item{method}{(string) - one of the above described methods}

\item{layout}{(boolean) - automatically read plate layout file}

\item{PLC}{(boolena) - automatically calculate pathlength correction}

\item{filename/pattern}{(string)}
}
\value{
data frame with reader data
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
Currently supported devices:
  BMG Readers (e.g. Omega ) - device='omega'
  Thermo Scientific (e.g. Varioskan) - device='varioskan'
Currently supported formats (method):
    SPabs : Single point absorption measurement (Omega, Varioskan)
    SPabsMatr: Single point absorption measurement in Matrix Layout format (Omega)
    SPfl:   Single point fluorescense measurement (Varioskan)
    KINabs: Kinetic absorption measurement (Varioskan)
Pathlength correction (PLC) is currently based on height and volume of a well/cylinder formula
}
\details{
importReaderData
}
\note{
todo: debugging, runtime optimization, error handling

filename patterns e.g.
}
\examples{
kin_df <- importReaderData("_KINabs_245_265_", method="KINabs", device='varioskan', layout=FALSE, PLC=TRUE)
}
\keyword{plate}
\keyword{readers}

