% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadPlateLayout}
\alias{loadPlateLayout}
\title{Reading Plate Layout Information from plate layout file}
\usage{
loadPlateLayout(barcode = "0000")
}
\arguments{
\item{reader_df,}{barcode="0000", setValueNA=TRUE, setSlopeNA=FALSE}
}
\description{
loadPlateLayout
}
\examples{
addPlateLayout(reader_df, barcode="0000", setValueNA=TRUE, setSlopeNA=FALSE)
}
\keyword{plate}
\keyword{readers}

