% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omegaSPabs}
\alias{omegaSPabs}
\title{BMG Omega: Singlepoint absorption measurement with multiple wavelengths - new file format (lists)}
\usage{
omegaSPabs(filename_list)
}
\arguments{
\item{filename_list}{}
}
\description{
omegaSPabs
}
\note{
num_wells needs to be soft-coded
}
\examples{
omegaSPabs(filename_list)
}
\keyword{plate}
\keyword{readers}

