% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omegaSPabsMatr}
\alias{omegaSPabsMatr}
\title{BMG Omega: Singlepoint absorption measurement with multiple wavelengths - old file formats (matrix)}
\usage{
omegaSPabsMatr(filename_list)
}
\arguments{
\item{filename_list}{}
}
\description{
omegaSPabsMatr
}
\note{
todo: temperature, shaking, num flashes
}
\examples{
omegaSPabsMatr(filename_list)
}
\keyword{plate}
\keyword{readers}

