% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pathlengthCorrectionFactor}
\alias{pathlengthCorrectionFactor}
\title{generating a simple pathlenght correction factor}
\usage{
pathlengthCorrectionFactor(liquid_volume = 0.2, well_radius = 0.34)
}
\arguments{
\item{liquid_volume=0.2,}{well_radius=0.34}
}
\description{
generating a simple pathlenght correction factor based on volume and radius of a cylinder
}
\details{
pathlengthCorrectionFactor
}
\examples{
pl_factor <- pathlengthCorrectionFactor(liquid_volume=0.2, well_radius=0.34)
}
\keyword{plate}
\keyword{readers}

