% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{plateReader}
\alias{plateReader}
\alias{plateReader-package}
\title{A library for importing plate reader files}
\description{
A library for importing plate reader files
}
\details{
currently supported readers: Thermo Varioskan (and related readers), BMG Omega (and related BMG readers)
}
\note{
date: 141122

last modification: 141129
}
\author{
mark doerr \email{mark.doerr@uni-greifswald.de},
  and \email{mark@ismeralda.org}
}

