% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readerDF2Array}
\alias{readerDF2Array}
\title{converts DataFrame to an easy plottable 3D array}
\usage{
readerDF2Array(reader_data_df, num = 0, wavelength = 0, add_zero = TRUE,
  use_slope = FALSE)
}
\arguments{
\item{reader_data_df,}{num=0, wavelength=0, add_zero=TRUE, use_slope=FALSE}
}
\description{
readerDF2Array
}
\examples{
readerDF2Array(reader_data_df, num=0, wavelength=0, add_zero=TRUE, use_slope=FALSE)
}
\keyword{plate}
\keyword{readers}

