% Generated by roxygen2 (4.0.2): do not edit by hand
\name{varioskanKinAbs}
\alias{varioskanKinAbs}
\title{Thermo Varioskan: Kinetic absorption measurement with multiple wavelengths - new file format (lists)}
\usage{
varioskanKinAbs(filename)
}
\arguments{
\item{filename_list}{}
}
\description{
varioskanKinAbs
}
\note{
todo - adding number of measurements
}
\examples{
varioskanKinAbs(filename_list)
}
\keyword{plate}
\keyword{readers}

