% Generated by roxygen2 (4.0.2): do not edit by hand
\name{varioskanSPfl}
\alias{varioskanSPfl}
\title{Thermo Varioskan: reading all fluorescent measurements into one data frame - new file format (lists)}
\usage{
varioskanSPfl(filename, barcode = "0000")
}
\arguments{
\item{filename}{}
}
\description{
varioskanSPfl
}
\note{
todo - adding correct number of measurements
}
\examples{
varioskanSPfl(filename)
}
\keyword{plate}
\keyword{readers}

